replace old extras plug. leave spherical harmonics in its own extras file, but in the surface fitting menu
